Fix windows bug that generated unsafe file paths in the rule description
