+++ halogen/mfbot.py
@@ -1,14 +1,15 @@
 import argparse
 import glob
 import os
-from lib.parser import get_file
-from lib.generator import yara_image_rule_maker
-from lib.render import yara_print_rule
-
+from halogen.parser import get_file_hash
+from halogen.generator import yara_image_rule_maker
+from halogen.render import yara_print_rule
 
 
 class MFBot:
     """ Malicious File Bot Class """
+    get_file: tuple
+
     def __init__(self) -> None:
         args = MFBot.parse_args()
         self.yara_base_file = args.file
@@ -16,7 +17,7 @@ class MFBot:
         self.idat = args.idat
         self.jpgsos = args.jpgsos
         self.dir = args.dir
-        self.dirhash = []
+        self.dir_hash = []
         self.name = args.name
         project_info_data = """
 -- Halogen ------------------------------------
@@ -25,7 +26,7 @@ Authors: Wyatt Roersma, Kyle Eaton, Devin Smith
         """
 
     @staticmethod
-    def parse_args()-> iter:
+    def parse_args() -> iter:
         """ Parse any options passed to the the script """
         parser_args = argparse.ArgumentParser(description="Halogen:  Automatically create yara \
             rules based on images embedded in office documents.")
@@ -49,7 +50,7 @@ Authors: Wyatt Roersma, Kyle Eaton, Devin Smith
         """mfbot.run() is the core function to call that will return all information
         generated by mfbot.
         returns: rule_dict - dictionary of rules. """
-        self.get_file = get_file(self)
+        self.get_file = get_file_hash(self)
         rule_dict = yara_image_rule_maker(self)
         if rule_dict is not None:
             return rule_dict
@@ -62,14 +63,14 @@ Authors: Wyatt Roersma, Kyle Eaton, Devin Smith
     def dir_run(self):
         """ runs through the process with a directory instead of a single file.
         returns: combo list. """
-        filelist = glob.glob(self.dir + "/*")
+        file_list = glob.glob(self.dir + "/*")
         combo = []
-        for f in filelist:
-            if os.path.isfile(f):
+        for file_name in file_list:
+            if os.path.isfile(file_name):
                 self.image_name = None
-                self.yara_base_file = f
-                self.get_file = get_file(self)
-                self.dirhash.append(self.get_file[0])
+                self.yara_base_file = file_name
+                self.get_file = get_file_hash(self)
+                self.dir_hash.append(self.get_file[0])
                 rule_dict = yara_image_rule_maker(self)
                 if rule_dict is not None:
                     for i in rule_dict:
